Moved all the styling methods to just one class
